Implement Laskin using real numbers

+++ src/laskin/Laskin.java
@@ -2,40 +2,40 @@ package laskin;
 
 public class Laskin {
 
-    private static int tulos;  // Muuttuja tulokselle
+    private static double tulos;  // Muuttuja tulokselle
     
     public void nollaa() {     // Nollaa tuloksen
         tulos = 0;
     }
 
-    public int annaTulos() {
+    public double annaTulos() {
         return tulos;
     }
     
-    public void lisaa(int n) {
+    public void lisaa(double n) {
         tulos = tulos + n;
     }
 
-    public void vahenna(int n) {
+    public void vahenna(double n) {
         tulos = tulos - n;
     }
 
-    public void kerro(int n) {
+    public void kerro(double n) {
         tulos = tulos * n;
     }
 
-    public void jaa(int n) {
+    public void jaa(double n) {
         if (n == 0) throw new ArithmeticException("Division by zero");
         else tulos = tulos / n;
     }
 
-    public void nelio(int n) {
+    public void nelio(double n) {
         tulos = n * n;
     }
 
-    public void neliojuuri(int n) {
+    public void neliojuuri(double n) {
         if (n < 0) throw new ArithmeticException("Square root of negative number");
-        else tulos = Double.valueOf(Math.sqrt(n)).intValue();
+        else tulos = Math.sqrt(n);
     }
     
     public void virtaON() {
